+++ functions/email/index.js
@@ -1,8 +1,11 @@
 /* eslint-disable no-console */
 const fetch = require('isomorphic-fetch')
 const FormData = require('form-data')
+const { isEmpty } = require('lodash')
 
-module.exports.send = (config, { to, subject, html, confName }) => {
+module.exports.send = (config, {
+  to, contact, subject, html, confName,
+}) => {
   if (!config || !config.key || !config.domain) {
     return Promise.reject(new Error('Mailgun configuration mailgun.key or mailgun.domain not found.'))
   }
@@ -23,7 +26,10 @@ module.exports.send = (config, { to, subject, html, confName }) => {
   to.forEach((dest) => {
     if (dest) form.append('to', dest)
   })
-
+  const cc = !isEmpty(contact) && /\S+@\S+\.\S+/.test(contact) ? contact : null
+  if (cc) {
+    form.append('cc', cc)
+  }
   return fetch(endpoint, {
     headers: { Authorization: `Basic ${token}` },
     method: 'POST',