+++ src/screens/conference/login/login.jsx
@@ -1,16 +1,36 @@
-import React from 'react'
+import firebase from 'firebase/app'
+import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 
 import { LoadingIndicator } from 'components/loader'
 import IconLabel from 'components/iconLabel'
 import Button from 'components/button'
-import AuthErrorModal from './authError'
 import './login.css'
 
 const Login = ({
-  authenticated, providers, signin,
+  authenticated, initialized, providers, signin,
 }) => {
-  if (authenticated) return <LoadingIndicator className="login-loading" />
+  const [authenticating, setAuthenticating] = useState(false)
+  const [errorMessage, setErrorMessage] = useState()
+
+  useEffect(() => {
+    setAuthenticating(true)
+    firebase
+      .auth()
+      .getRedirectResult()
+      .then(() => {
+        setAuthenticating(false)
+      })
+      .catch((error) => {
+        setAuthenticating(false)
+        setErrorMessage(error.message)
+      })
+  }, [])
+
+  if (!initialized || authenticating || authenticated) {
+    return <LoadingIndicator className="login-loading" />
+  }
+
   return (
     <div className="login">
       <h1 className="login-title">Connexion</h1>
@@ -19,12 +39,13 @@ const Login = ({
           <IconLabel icon={`fa fa-${provider}`} label={`with ${provider}`} />
         </Button>
       ))}
-      <AuthErrorModal />
+      {errorMessage && <p className="login-error">{errorMessage}</p>}
     </div>
   )
 }
 
 Login.propTypes = {
+  initialized: PropTypes.bool,
   authenticated: PropTypes.bool,
   providers: PropTypes.arrayOf(PropTypes.string),
   signin: PropTypes.func.isRequired,
@@ -32,6 +53,7 @@ Login.propTypes = {
 
 Login.defaultProps = {
   providers: [],
+  initialized: false,
   authenticated: false,
 }